Update functions
Updated name search function so that list created will make copies of the details for each pokemon. Before, it would cut the details from the pokedex and leave empty boxes in its place.
Added a new function that will remove the current search list everytime a key is pressed so a new one can be made.